master_fiscovery_fkie: integrated pull request #24
Thanks for creating the PR to @garyservin and @mikeodr!

The change lets you define an interface by `~interface`, `ROS_IP` envar
or append the interface to multicast group like
226.0.0.0@192.168.101.10. The master_discovery then binds to the
specified interface and creates also an unicast interface for active
requests on communication problems or if `~robot_hosts` are defined.

Now you can also disable the multicast communication by setting
`~send_mcast` to false. In this case the requests are send to hosts
defined in `~robot_hosts`.