Support hub topology by adding favor_unicast param
In cases where multicast sending is enabled but favor_unicast
is also set, we only listen for multicast messages to detect
remote ROS masters and then use unicast to send out heartbeat
requests and state.  What this allows us to do is deploy multimaster
nodes as a hub rather than a mesh: master_discovery nodes with
multicast and favor_unicast enabled with act as a client to any
master_discovery nodes with multicast enabled but favor_unicast
disabled.

The one caveat is that this can only work if all ROS masters are
on a separate host.  If more than one master_discovery is on the
same host we have to use multicast to ensure that all processes
receive the heartbeat state/request messages.  In this case we
default back to the 'normal' mesh like behavior.

This is useful for situations where:
1) Each ROS master is running on a separate host
2) Only one ROS master needs to talk to every other ROS master
   (hub)
3) We do not know beforehand the name or IP of the hub and
   want hub clients to attach to the hub automatically
   (otherwise we could disable multicast on each hub client
   and set the hub IP in the robot hosts param).